:feat Add/Remove friends with Queries

+++ src/main/java/com/example/Player/dto/FriendDTO.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class FriendDTO {
-    private int friendId;
+    private long friendId;
     private String friendName;
 
-    public FriendDTO(int friendId, String friendName) {
+    public FriendDTO(long friendId, String friendName) {
         this.friendId = friendId;
         this.friendName = friendName;
     }
